refactor : AdminController 제거
nextstep과 auth 패키지 간의 일방향 의존을 위해 Theme/ScheduleController로 Theme/Schedule Admin API 이동
